Attempting to make image.cpp slightly more consistent

+++ cplusplussource/image.h
@@ -23,7 +23,7 @@ struct reservation {
         int endBlock;
 };
 
-void setPixel(int x, int y, unsigned char color);
+void setPixel(uint32_t x, uint32_t y, unsigned char color);
 void drawRect(int x, int y, int width, int height, unsigned char color);
 void initializeImage();
 void drawCharacter(int x, int y, char c);
@@ -31,8 +31,8 @@ void drawString(int x, int y, std::string str);
 void drawFancyString(std::string str, int16_t x, int16_t y);
 uint16_t getTextWidth(std::string str); 
 bool drawCenteredString(std::string str, int16_t y);
-int getQrCodeSize(std::string str);
-int putQrCode(int x, int y, std::string str, int scale);
+uint16_t getQrCodeSize(std::string str);
+uint16_t putQrCode(uint16_t x, uint16_t y, std::string str, uint16_t scale);
 unsigned char reverseByte(unsigned char x);
 void invert();
 void rotate();
